Move github-specific mocks to workflows folder
